requirements.txt
tensorflow
keras
flask
numpy
streamlit
pillow
opencv-python